Install sqlite 3.44.5 source files
This installs sqlite 3.44.5 source files.  It does not include the new
version of sqlite in the build.  The actual version of sqlite that is
installed is controlled by a trunk-stable build flag, which is not
modified in this CL.

Note: the installation on main is primarily to create a CL that can be
cherry-picked to older releases.  The current plan of record is that
main will adopt 3.50.x.

Flag: EXEMPT source install only
Bug: 430889718
Bug: 281064726
Test: presubmit
Cherrypick-From: https://googleplex-android-review.googlesource.com/q/commit:53259b30e767bf8e7e9f61740d33c1c33462a7ff
Merged-In: I805ccd030564c96992b00dc8c8a6b12eb936458c
Change-Id: I805ccd030564c96992b00dc8c8a6b12eb936458c